django/products/requirements.txt
Django==2.2.24
djangorestframework==3.9.4
gunicorn==19.9.0
requests==2.22.0
django-rest-swagger==2.2.0
celery==4.3.0
psycopg2-binary==2.8.5
pytz==2024.1
psycopg2==2.8.5
